Refactor dataset code in response to code review
